api_eventos/requirements.txt
Flask>=2.0
Flask_SQLAlchemy>=3.0
python-dotenv>=1.0
psycopg2-binary>=2.9
